fixed multi-author/narrator in the OPF

+++ myx_utilities.py
@@ -296,10 +296,10 @@ def createOPF(book, path):
         template = file.read()
 
     # - Author -
-    if book.getAuthors() == '__unknown__':
-        template = re.sub(r"__AUTHOR__", '', template)
-    else:
-        template = re.sub(r"__AUTHOR__", book.getAuthors(), template)
+    authors=""
+    for author in book.authors:
+        authors += f"\t<dc:creator opf:role='aut'>{author.name}</dc:creator>\n"
+    template = re.sub(r"__AUTHORS__", authors, template)
 
     # - Title -
     template = re.sub(r"__TITLE__", book.title, template)
@@ -308,22 +308,34 @@ def createOPF(book, path):
     template = re.sub(r"__SUBTITLE__", book.subtitle, template)
 
     # - Narrator -
-    template = re.sub(r"__NARRATOR__", book.getNarrators(), template)
+    narrators=""
+    for narrator in book.narrators:
+        narrators += f"\t<dc:creator opf:role='nrt'>{narrator.name}</dc:creator>\n"
+    template = re.sub(r"__NARRATORS__", narrators, template)
 
     # - ASIN -
     template = re.sub(r"__ASIN__", book.asin, template)
 
     # - Series -
-    if len(book.series) and len(book.series[0].name):
-        template = re.sub(r"__SERIES__", book.series[0].name, template)
-        template = re.sub(r"__SERIESPART__", book.series[0].part, template)
-    else:
-        template = re.sub(r"__SERIES__", "", template)
-        template = re.sub(r"__SERIESPART__", "", template)
+    series=""
+    for s in book.series:
+        series += f"\t<ns0:meta name='calibre:series' content='{s.name}' />\n"
+        series += f"\t<ns0:meta name='calibre:series_index' content='{s.part}' />\n"
+    template = re.sub(r"__SERIES__", series, template)
+    
+    # if len(book.series) and len(book.series[0].name):
+    #     template = re.sub(r"__SERIES__", book.series[0].name, template)
+    #     template = re.sub(r"__SERIESPART__", book.series[0].part, template)
+    # else:
+    #     template = re.sub(r"__SERIES__", "", template)
+    #     template = re.sub(r"__SERIESPART__", "", template)
 
-    opfFile=os.path.join(path, "metadata.opf")
-    with open(opfFile, mode='w', encoding='utf-8') as file:
-        file.write(template)
+    if (myx_args.params.dry_run):
+        pprint (template)
+    else:
+        opfFile=os.path.join(path, "metadata.opf")
+        with open(opfFile, mode='w', encoding='utf-8') as file:
+            file.write(template)
 
     return
